Feat(norms): add chainability of specifications
While specifications for the same generic type can be combined, as well
as specification results, chainability provides a more convenient and
fluid syntax. This becomes especially handy when it is necessary to
evaluate and combine multiple specification for different types with
a variety of instances and merge everything into a single result.

The implementation tries hard to provide a full type safe usage
guarantees. Under the hood it needs to do some "unexpected" bulking to
work around the weaknesses of the Kotlin/Java type system.

RISDEV-1680