Refactor: rework chips components
This affects:

- Chips list
- Chips input
- Keywords chips

These use shared logic and very similar in terms of implementation and
interaction (even though there are some layout and requirement details
that make it more convenient to keep them separate). Note that the date
chips input is not affected, since it's completely separate from the
others.

This commit:

- Changes the list and chips v-models to be immutable according to the
  new immutable v-model pattern (see ADR 15)
- Also refactors the focus handling to be a v-model to make it easier to
  use and less verbose
- Fixes a bunch of bugs in the chips and components using the chips
  along the way
- Improves test coverage for all 3 components, splits the tests for
  keywords and regular chips into 2 files

RISDEV-2038