Chore(norms): fix fitness tests
The architecture fitness tests were not properly working. That means
some of the rules encoded as fitness tests were not properly enforced
and some breaches were made.
There is the problem that our rules apply on how the code "looks" like
in Kotlin. But the tests work only on the pre-compiled Java code. The
Kotlin compiler does create a lot of helper classes in Java to represent
the Kotlin code. E.g. static classes for global functions. These need to
be excluded for some fitness tests as they cause failures were no rule
was broken (in Kotlin).
Unfortunately we applied such a filter for Kotlin-specific-classes on
a global level. And this filter was just too aggressive and filtered out
many parts which are actually very relevant for many rules. Thereby some
code sections are no more compliant with the rules because they have
been filtered out/ignored.

As solution, this filter gets applied more selectively for specific
tests. Mostly naming convention tests. All other tests, especially
dependency tests, act on all source classes again.
But there is now some code that is not compliant. The old issue withe
the date encoding functions has been fixed in place. But the bigger
issues have been left there. BUT they are explicitly marked as
exceptions within the fitness tests.
It is the goal to achieve consense in the team again and resolve the
issues or adapt the architecture and thereby the fitness functions.

RISDEV-0000