Feat(norms): use collections for norm entity
Within our domain model, we use exclusively property values (`order`) to
determine the ordering of elements in a listing. This approach has some
advantages. Though these are up for discussion again.
In any case, the usage of the `List` class for these introduces a second
ordering on top of the elements by their index. That might lead to
confusion or is simply not very expressive and clean code.
The alternative is to use the interface type `Collection` instead. This
type does not allow to rely on any ordering for this listing. Under the
hood, we can still instantiate `List`s to construct these parameters.
But within the entity, the type is reduced to the capabilities of
a `Collection`. Alternative `Collection` implementations are a `Set`.
But `Set`s are much less computation efficient. So we don't wanna use
them for now. We might resolve that issue in future and actually use
`Set`s.

RISDEV-2315