Fix: WIP adjust components, tests to updated types
This fixes all TypeScript complaints resulting from changes made in the
previous commit.

⚠️ Fair warning: This is a WIP to hand over some work I started, and in
all likelihood *badly broken*. Among other things it is entirely
untested as required changes are not yet implemented in the backend. It
also assumes that documentation of a norm contains *only* document
sections on the first level, and *only* articles on the second level.
Something similar used to be the case, so the current implementation is
an MVP of sorts, but it's no longer true for the new data structure and
needs to be updated soon.

RISDEV-2315