Fix: updating of validation errors with undefined
RISDEV-2271
Errors, that occur in the input group in dependency to the values of other input fields, have been always overriden by an undefined value, because the input value itself was validated and correct. Removing the emit with undefined, after the user completed the date solves this issue and was redundant anyway. Because the only use case that we would need the input field to emit undefined is, if the value was incorrect before und is corrected by the user. In this case the user needs to delete the wrong value, the backspaceDelete will fire and resets the validation error to undefined. With this solution, resetting the error on the input itself still works and it does not override errors on a higher level, that only occur in collision with other input values.