Feat(norms): adapt database adapter documentation
This applies the changes of the domain model for the documentation part
to the database schema and all the work the adapter has to do as
translation.

We decided to jump over the first iteration of this domain model
extension. We can re-use the former used `articles` and `paragraphs`
tables which still exist. So a diff to the state before the last
iteration is much smaller and simpler.

RISDEV-2105