Chore(compose): increase database init delay
When the database service in the local Docker compose setup gets
started, it does two initialization steps. First something based on the
regular entry point that starts the database service inside the
container and creates database, users, etc. based on
environment variables. The second manual step by us is to install some
extension by using a command that connects to that database service. It
runs with a delay. This is a hope-for-the-best approach and requires the
service to finish within that delay.
The first step only happens on the first run with a fresh and empty
database. In that case, the first step takes some time to finish. The
second step depends on that one, so if it is not completed yet, it
fails. Symptoms are errors like socket not available or database does
not exist.
For runs without the first step, the delay was always enough. With, it
very often is not. The assumption is that the second step was introduced
without testing it with a clear environment. As most developers locally
have the database already setup and clean it rarely, this has not been
discovered for some time.

The solution is still the same approach as before, just increasing the
delay to a high enough value that increases the probability enough that
the first initializations steps finished.

RISDEV-0000