Chore(norms): improve domain DSL for tests
First of all it must be able to create metadata sections with `null` as
the value for child sections. That required some tweaking for the
builders.
Moreover some little stronger naming was established, plus some access
to low level builders in the DSL (i.e. constructor metadata sections
outside of a norm).

RISDEV-0000