WIP
Removes NormenKomplex page and from the menu and adapts the e2e tests for that.
Adds a border bottom to the last item in the documentsection loop
Adapts the open/close icons for the Bestand page expandable component.
Adjusts the tableOfContents.spec.ts
Update the router.replace to the content page when importing a zip.

RISDEV-2335