Feat(norms): extend empty data detection
The function to detect empty data structures was extracted to test it
independently of the component using it.
In general this should be more of a business rule. But at the moment we
don't have the data management and flow to apply them.

This also extends the function to report blank strings as empty and any
number as not empty.

RISDEV-2120