Correctly update search input model
RISDEV-2324
As the search now sends the search query as uri parameters, the documentUnitSearchInput model could be simplified to have only strings, no objects anymore. The url parameters are passed on reload to the search input form to keep the state of the search, a deep watcher was added for that purpose.