Fix: downgrade spring k8s config dependency
The former upgrade was performed due to a high security vulnerability.
Unfortunately was that a full major version update that currently
doesn't work on our platform.
Turned out, pinning only the affected child-dependency is closing the
vulnerability too. Thereby the upgrade of the spring k8s config is
postponed.

RISDEV-0000