Refactor: remove unnecessary imports from tests
Everything vitest related is exposed globally and doesn't need to be
imported manually.

Currently SonarCloud reports a bunch of codesmells related to the tests,
where it wrongly claims that no assertion has been made. We suspect this
is because it gets confused about which testing library we use. Cleaning
up imports might fix that.

RISDEV-0000