Fix overriding of local value
RISDEV-2857
When the noClear prop is set, that means, that it is technically not possible to 'delete' a value via keyboard. We just want to have a local state of no input, until I select a new input or leave the field. That's wh we need to prevent emitting an undefined value or let the props override the local value, when the input text is empty.