Fix pageable in related docunit search
RISDEV-2957
The search was not implement properly with our pageable component (always returned the results for page 0). The loading spinner was moved to the search result list, so we do not need to repeat it in each component, that implements it. Also we sync the different states better here (empty, loading, etc. ...).