Improve date unknown behaviour
RISDEV-2987
Add fix for isEmpty behaviour - an entry with checkbox 'anhängig' checked, should still be empty. It should not be possible to add an entry with only 'anhängig' clicked. Therefore 'pending' was removed from the fields to be iterated over, when checking for an empty entry. Also the check regarding the missing required field logic was moved from the isEmpty method, to the getMissingRequiredField method respectively. The behavior of the visibility of the decision date field was adapted for more consistency in previous and ensuing decision (the field is not only disabled, but not visible anymore, when the date is unknown).