Add tailwind preflight for list styles
RISDEV-3678

Per default, in tailwind preflight lists are unstyled. We used to override that behaviour in the TextEditor css (which by the way overrode the list styles globally, because it was not scoped). Setting the list style to auto in the TextEditor had the effect, that the list style was set to the browser default (which is 'disc' for unordered lists and 'decimal' for ordered lists). Meaning, that even if the list had no list-style attribute, it was set to the default list style, which had bad sideeffects (see RISDEV-3678, here docx Absatzformate (e.g. headlines) create lists with no list styles → after setting it to the default headlines had decimal indexes for example).

We solve this bis not overriding the preflight’s list style, because it is overriden anyways by inline styles, if given in the docx. So normal lists should work the same, only lists with no list styles fall back to the preflights list style.

We still need to override the preflights list paddings and margins because per default, we want lists to be indented, thus - use the browsers default margin and paddings for lists. We do this globally, by adding a preflight extension to the global.scss.

Accordingly we needed to update also the ChipsList, because here we want no margins and paddings in this special case.

Also the procedure combobox was fixed, it should not be clearable.