Refactor code to always search on button click
RISDEV-4003
In order to avoid having too many conditionals and variables (e.g. isIdenticalSearch), this solution decouples the search completely from the route. The route is still pushed to the router, in order to maintain the search on reload, but now the search is always triggered, when the search button is clicked and on first mount.