Fix: adding from search
RISDEV-4011
When searching without any search parameters and adding an item from that search, the entry was directly deleted. That was because of the cleaning mechanism in onBeforeUnmount. We created a new copy from that search entry and emitted the value, but at the same time, we were watching the local copy of the entry, which was still empty. Therefore the entry got removed on unmount. Now we are saving the search result copy also in the local modelvalue, so the cleaning mechanism does not trigger. This was fixed for active citations, ensuing decisions and previous decisions.