Enable repeated search in DocumentUnitSearchEntryForm (#1525)
* Add repeated search in DocumentUnitSearchEntryForm

The test is WIP.
RISDEV-4003

* Refactor code to always search on button click

RISDEV-4003
In order to avoid having too many conditionals and variables (e.g. isIdenticalSearch), this solution decouples the search completely from the route. The route is still pushed to the router, in order to maintain the search on reload, but now the search is always triggered, when the search button is clicked and on first mount.

* Fix search on changed route and test

This search has been triggered twice
RISDEV-4003

---------

Co-authored-by: Leonie Koch <leonie.koch@digitalservice.bund.de>