Updating editable list to work with unique id
RISDEV-4316

Relying on indexes of a list conflicts when changes are received from a remote user simultaneously. With this update editing list items are referenced by their unique IDs (UUIDs). There is no need to remove an empty entry as it is not added to the list if it is empty. This prevents a newly created, unsaved item from being merged into the list.