Treat non-list lists as paragraphs
Lists that have a numId "0" are not shown as lists in word, and therefore instead of just hiding the numbering symbol we are now treating them as paragraphs instead. This is in line with how word displays them. Additionally, if there is no matching list level in the abstract numbering definition, we also treat it as a paragraph.

RISDEV-4114