Deal with hanging indents
Hanging indentation, which specifies the indentation to be removed from the first line, is now subtracted from the left indentation if it is present. This ensures that paragraphs where the first line is not indented at all (left indent and hanging indent are the same value) do not get a margin left attached to them. If the values are not equal, the paragraph will still be indented.

RISDEV-4447