Fix flaky docunit handover e2e util
RISDEV-3973
In some cases the PUT request to hand over the doc unit via email is not executed. It is unclear why. Waiting for the XML preview to be loaded (GET request) on mount of the handover component fixes the problem. Might be competing requests or just a timing problem.