Link decision from search after validation
RISDEV-4560
The decision from search must be assigned only after the validation for required fields ran through. Because assigning the decision before validation, has the effect, that the decision input fields are set to readonly. In case the validation fails, we can not adjust these inputs anymore. Also assignging the legalPeriodical to the current references from the store is now handled reactively.