Shorten undefined/empty check in deviating data
RISDEV-4371
Shorten the following expression modelValue.deviatingEclis && modelValue.deviatingEclis.length > 0 via optional chaining does not work, because when we use modelValue.deviatingCourts?.length, if modelValue.deviatingCourts is undefined or null, the expression evaluates to undefined, not false. However, Vue expects a boolean in the binding, and undefined is not strictly false.

To ensure a proper boolean value, we can explicitly convert the result into a boolean using Using double negation !! (forcing a boolean conversion):
- !!undefined becomes false.
- If modelValue.deviatingCourts exists and its .length is greater than 0, it evaluates to true.
- If modelValue.deviatingCourts is undefined or has a length of 0, it evaluates to false.