Fix bug RISDEV-4730
RISDEV-4730
When clicking on cancel on a new norm reference entry, that haven’t been saved yet, the singleNorm entries vanish, but the display value of the norm abbreviation stayed. The model value updates were not registered, because the modelvalue was undefined before and after cancelEdit is again undefined. A deep watcher solves this issue because it tracks changes for all nested properties inside an object. So, when norm.value.normAbbreviation is set to undefined, the watcher is triggered because it now deeply observes every property inside norm.value. Also we need to manually set a new empty single norm entry, because it otherwise does not seem to work.