Apply code review remarks & fix Rd-Link validation
For better readability we decided to avoid loops in the  texts components even if that means some code duplication.

The validation of border number links (Randnummernverlinkung) was previously only done in "Titelzeile" and "Leitsatz". This has been fixed as it is theoretically possible to add border number links to every text category. This is the reason why we now validate every text category (except from decisionName where adding links is not possible).

RISDEV-4729