Improve search query performance (#2391)
RISDEV-0000
These additional categories are not relevant for the search results but are still fetched for each row (N+1 problem) because of the EAGER fetching. With lazy fetching, they will only be queried on demand (e.g. when loading the whole doc unit).

For each row that is fetching during a search, we do 9 SQL queries. So for fetching a search that has 100 results, we run 900 SQL queries against the database. With this change, we reduce the number of queries by 5 per row. Why it does not just join these additional tables in the first place is unclear to me.