Cancel+debounce Combobox requests + delayed enter (#2426)
RISDEV-5663

* Cancel+debounce Combobox requests + delayed enter
Previously: Requests on every keystroke. Now, we debounce and cancel requests. Also, when enter is pressed while a search is ongoing, we defer the enter to select the first item to when the request resolves.

* Fix focus logic
Focusing was broken because keeping track of the html element list of the combobox options relies on the correct ordering which conflicts with vue's reusage of html elements. For now, we need to accept the negative performance impact.