Replace existing sources on update
RISDEV-6888
So far, the current source (domain model) would get added to the sources list of the DTO every time the doc unit was saved. Now a new source is only added to an empty sources array. Otherwise, the latest existing source will be overwritten if the value has changed.