Fix saving of doc unit with reference source (#2624)
RISDEV-6888
A doc unit has direct relationships to its Fundstellen (references) and potentially an indirect relationship via the Quelle (source) if it was created via Periodikumsauswertung (edition). When the same Fundstelle was linked to the doc unit via both relationsships, the doc unit could not be saved because both relationships had full cascading which lead JPA to try to update the same entity (Fundstelle) twice. To resolve this, the cascading merge was removed from the source.

So far, the current source (domain model) would get added to the sources list of the DTO every time the doc unit was saved. Now a new source is only added to an empty sources array. Otherwise, the latest existing source will be overwritten if the value has changed.