Globally stub InputMask with InputText in tests
RISDEV-7321
InputMask evaluates the cursor position on every keystroke using a library that calculates the exact pixel offset. Determining the cursor position in the browser is complex and requires additional libraries. However, our browser environment for Vitest does not implement any layout-related functionality, meaning the required functions for cursor offset calculation are missing.

To fix this, we would either need to manually implement/mock these missing functions or switch to a different browser simulation library—both of which are too much effort.

Since our unit tests do not aim to verify the input masking itself but rather the actual validation of a date input (such as whether the input is complete and valid), we only need to handle the input as a string. A simple text input is sufficient for this purpose.

Therefore, we globally stub InputMask with InputText for all unit tests.