Parse patch JSON for assertions
RISDEV-7247
Instead of asserting on large JSON strings, we parse the JSON and assert on the properties. This allows to ignore dynamic dates, e.g., lastUpdatedAt.
This has only been done for one method. Before refactoring the rest, we should check whether this is the best approach and what parts can be extracted to helper methods (e.g. the lastUpdate events, they are always the same)