Add metadata to management data (#2694)
RISDEV-7401

* Add metadata to management data
* Fix name of ManagementDataTransformerTest
* Move FE components to management-data folder
* LocalDate -> Instant
* Cascading for ManagementData
* lastUpdate attributes in PATCH
* Add user to transform
* Safe managementData for new DocumentationUnit
* Set management data when uploading/deleting file
* Respond with createdByUserName on /new

The user was missing so that the transformed docunit response would not contain the username.

Instead of asserting on large JSON strings, we parse the JSON and assert on the properties. This allows to ignore dynamic dates, e.g., lastUpdatedAt.
This has only been done for one method. Before refactoring the rest, we should check whether this is the best approach and what parts can be extracted to helper methods (e.g. the lastUpdate events, they are always the same)

Co-authored-by: anne-ds <anne.schumann@digitalservice.bund.de>