Reuse user group service method
RISDEV-9143

This is a more restrictive approach to handling a disambiguated user group, which we are currently using in our user API. We could adjust the method to be more flexible by adding support for the longest segments. Finally, we should stream all user groups through this once we discuss our user group structure into account. If user is not allowed to be under two categories, an error is helpful.

* User Group fetching in service
* User controller test fix
* Fix code smells
* Add user group tests