Restructure frontend components
RISDEV-8883
We have now moved the bulk assign logic into a seperate component, so we keep the ResultList a High-Level 'View' component, that renders data. The new 'smart' BulkAssignProcessStep is the logic component, that encapsulate the bulk assign use case (and can be enriched with furhter bulk assign actions) and a dumb UI component AssignProcessStep, that just fetches process steps and users to let the user choose the next process step. The code has been refactored to align with the BulkAssignProcedure component.