Clear process steps when assigning new docoffice
RISDEV-9167
dquote> This logic applies when a document's Documentation Office is reassigned.

scenario 1: docoffice A assinging to docoffice B -> all previous process steps are cleared. This action deletes all underlying documentationUnitProcessStep data. For our history logs the loss of the documentationUnitProcessSteps is acceptable, because we only need these documentationUnitProcessSteps to retrieve the user data dynamically. As the process step change the history logs save a description like 'Schritt geändert: [from] -> [to]', which is hardcoded information in the description field and will therefore be there even though the underlying information is not there anymore. Only the user data can not be retrieved, but in this case it is fine, because the docoffice B is not allowed to see user information from docoffice A anyways, they will only see 'Person geändert'.

scenario 2: docoffice B assigns docunit back to docoffice A -> when assigning BACK to docoffice A, this information can not be restored, because the underlying documentationUnitProcessStep data is gone. Meaning, that we still do only show 'Person geändert', even though the user from docoffice A should be able to see the user information. We took this tradeoff, because this is a very rare edge case and to properly handle previous process steps for other docoffices after reassigning, we need to store the docoffice information seperately in the documentationUnitProcessStep and not in the user of the documentationUnitProcessStep. (Because the user will not always be there to provide that information).

This trade-off was made because this 'reassign back' scenario is a very rare edge case. The proper solution would require storing the documentation office information separately within the DocumentationUnitProcessStep entity to preserve history, which was deemed too much effort for this very edge case.