Refactor itemService to work with new autocomplete
This improved the type safety by no longer combining the types of all
the possible return values into one but instead specifying the type for
every endpoint.

We also placed all the formating logic into only one place as this was
previously spread acrose both the comboboxItemService and the components
using it.

This also simplified using the item service when not wanting to use
ComboboxItem but needing the actual values.

RISDEV-10454